fix 1 for marcview

+++ automatic/marcview/tools/chocolateyinstall.ps1
@@ -20,8 +20,25 @@ Remove-Item $fileLocation -ea 0 -force
 # The installer places the Start Menu shortcut in the current user profile only
 $UserStart = Join-Path $env:APPDATA '\Microsoft\Windows\Start Menu\Programs'
 $OrigLNK = Get-ChildItem -Path $UserStart -Filter "$env:ChocolateyPackageName.lnk" -Recurse
+
 # Creating a Start Menu shortcut for all users.
-$AllStart = Join-Path $env:ProgramData '\Microsoft\Windows\Start Menu\Programs'
-$NewLNK = Copy-Item $OrigLNK.FullName $AllStart -Force -PassThru
-$null = Remove-Item $OrigLNK.FullName
+$StartPrograms = Join-Path $env:ProgramData '\Microsoft\Windows\Start Menu\Programs'
+If (Test-Path $OrigLNK) {
+   $NewLNK = Copy-Item $OrigLNK.FullName $StartPrograms -Force -PassThru
+   $null = Remove-Item $OrigLNK.FullName
+} else {
+   # Installing as SYSTEM doesn't create a Start Menu shortcut at all
+   [array]$key = Get-UninstallRegistryKey -SoftwareName $packageArgs['softwareName']
+
+   if ($key.Count -eq 1) {
+      $InstallDir = Split-Path $key[0].UninstallString
+      $targetPath = Join-Path $InstallDir "MARCView.exe"
+      $StartPrograms = Join-Path $env:ProgramData '\Microsoft\Windows\Start Menu\Programs'
+      $shortcutFilePath = Join-Path $StartPrograms 'MARCView.lnk'
+
+      Install-ChocolateyShortcut -shortcutFilePath $shortcutFilePath -targetPath $targetPath
+   } else {
+     Write-Warning "Start Menu shortcut could not be created."
+   }
+}
